realms work in progress

+++ Source/ACE.Server/Physics/PhysicsObj.cs
@@ -2067,7 +2067,7 @@ namespace ACE.Server.Physics
             }
         }
 
-        public void calc_cross_cells(byte? instance = null)
+        public void calc_cross_cells(uint? instance = null)
         {
             CellArray.SetDynamic();
